requirements.txt
streamlit
openai
langchain
faiss-cpu
sentence-transformers
sympy
paddleocr
pytesseract
openai-whisper
python-dotenv
numpy
Pillow
